requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
Django==5.2.12
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.3
nltk==3.9.4
numpy==2.4.3
PyJWT==2.12.1
regex==2026.2.28
scikit-learn==1.8.0
scipy==1.17.1
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.3
tzdata==2025.3
